+++ src/Cove.Api/Program.cs
@@ -329,6 +329,9 @@ try
     builder.Services.AddScoped<PerformerScrapeService>();
     builder.Services.AddScoped<ScrapeAttemptService>();
     builder.Services.AddSingleton<DownloaderService>();
+    builder.Services.AddSingleton<IDownloaderService>(provider => provider.GetRequiredService<DownloaderService>());
+    builder.Services.AddSingleton<IDownloaderSiteLoginProvider, DownloaderSiteLoginProvider>();
+    builder.Services.AddScoped<IVideoFileMaintenanceService, VideoFileMaintenanceService>();
     builder.Services.AddSingleton<ITranscodeService, TranscodeService>();
     builder.Services.AddScoped<StashMigrationService>();
     builder.Services.AddScoped<ITagProvenanceService, TagProvenanceService>();